contact storage in one file
