--- src/adam
@@ -1,1 +1,0 @@
-adamgecko1234@gmail.com
