--- src/dad
@@ -1,1 +1,0 @@
-joshua.goodman@gmail.com
