--- src/jared
@@ -1,1 +1,0 @@
-jaredgoodman03@gmail.com
