--- src/mom
@@ -1,1 +1,0 @@
-ericagoodwoman@yahoo.com
